Add fully qualified names for std::string and std::map
The header TJAlienCredentials.h polluted the global namespace, included usually with CCDB
or something related.

Many source files in O2 took advantage of this.
This commit prepares for the removal of the polluted namespace.
The libraries/repo with TJAlienCredentials is JAliEn-ROOT and libjalieno2.
The problem was first noticed by @vkucera.