Add residuals for external detectors unless --skip-ext-det-residuals is passed
The unbinned residuals got channel data member which holds ITS chipID, TRD chamber ID
or TOF pad within the sector resp. The convention for non-TPC rows:
160-165: TRD layers
170:     TOF
180-186: ITS layers
The getters isTPC(), isTRD(), isTOF() and isITS() can be used to query the residual type
and getDetID() will return 0,1,2 and 3 for ITS, TPC, TRD and TOF points resp.
Also, getX() and getAlpha() will provide the tracking frame X and alpha of the residual.
These getters, when acting on the non-TPC residuals require initialized geometry helpers.
The initialization can be done manually via static method
o2::tpc::UnbinnedResid::init(long timeStamp = -1)
which will load from the CCDB the geometry corresponding to the requested timestamp (-1 : now).
If the initialization was not done in advance then it will be done internally at the 1st
query of getX() or getAlpha(), with the timestamp = -1.